reset base to fsl default

+++ scripts/submod_reset.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 #Cleans and resets a git repo and its submodules
 #https://gist.github.com/nicktoumpelis/11214362
-git reset --hard
+
+[ "$1" = "all" ] && git reset --hard
 git submodule sync --recursive
 git submodule update --init --force --recursive
-git clean -ffdx
+[ "$1" = "all" ] && git clean -ffdx
 git submodule foreach --recursive git clean -ffdx
